requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
crcmod-plus==2.3.1
cryptography==46.0.3
curl_cffi==0.13.0
cycler==0.12.1
dashscope==1.25.5
distro==1.9.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
firecrawl-py==4.13.4
fonttools==4.61.1
frozendict==2.4.7
frozenlist==1.8.0
google-api-core==2.28.1
google-auth==2.47.0
google-cloud-aiplatform==1.135.0
google-cloud-bigquery==3.40.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.16.0
google-cloud-resource-manager==1.16.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-genai==1.61.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.8
multidict==6.7.0
multitasking==0.0.12
nest-asyncio==1.6.0
numpy==2.4.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
peewee==3.19.0
pillow==12.0.0
platformdirs==4.5.1
playwright==1.57.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
pyflakes==3.4.0
pyparsing==3.3.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
websocket-client==1.9.0
websockets==15.0.1
yarl==1.22.0
